backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Pillow==10.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==3.1.3
